refactor: several minor updates on preset plugins
