+++ web/src/beta/features/PluginPlayground/Plugins/index.tsx
@@ -2,9 +2,10 @@ import { Collapse, Typography, Button } from "@reearth/beta/lib/reearth-ui";
 import { EntryItem } from "@reearth/beta/ui/components";
 import { useT } from "@reearth/services/i18n";
 import { styled } from "@reearth/services/theme";
-import { FC, useMemo, useState } from "react";
+import { FC, useState } from "react";
 
 import FileListItem from "./FileListItem";
+import useTitles from "./presets/useTitles";
 import usePlugins from "./usePlugins";
 
 type UsePluginsReturn = Pick<
@@ -47,66 +48,7 @@ const Plugins: FC<Props> = ({
     encodeAndSharePlugin(selectedPlugin.id);
   };
 
-  const categoryTitles: Record<string, string> = useMemo(() => {
-    return {
-      custom: t("Custom"),
-      ui: t("User Interface"),
-      communication: t("Communication"),
-      viewerScene: t("Viewer & Scene Settings"),
-      layers: t("Manage Layer"),
-      layerStyles: t("Manage Layer Style"),
-      camera: t("Camera"),
-      timeline: t("Timeline"),
-      dataStorage: t("Data Storage"),
-      extension: t("Extension")
-    };
-  }, [t]);
-
-  const pluginTitles: Record<string, string> = useMemo(() => {
-    return {
-      "my-plugin": t("My Plugin"),
-      "responsive-panel": t("Responsive Panel"),
-      sidebar: t("Sidebar"),
-      header: t("Header"),
-      "popup-plugin": t("Popup"),
-      "modal-window": t("Modal Window"),
-      "ui-extension-messenger": t("UI Extension Messenger"),
-      "extension-to-extension-messenger": t("Extension To Extension Messenger"),
-      "enable-shadow-style": t("Enable Shadow Style"),
-      "enable-terrain": t("Enable Terrain"),
-      "show-label": t("Show Label"),
-      "take-screenshot": t("Take Screenshot"),
-      "mouse-events": t("Mouse Events"),
-      "add-geojson": t("Add Geojson"),
-      "add-csv": t("Add CSV"),
-      "add-kml": t("Add KML"),
-      "add-wms": t("Add WMS"),
-      "add-czml": t("Add CZML"),
-      "add-3d-tiles": t("Add 3D Tiles"),
-      "add-google-photorealistic-3d-tiles": t(
-        "Add Google Photorealistic 3D Tiles"
-      ),
-      "add-osm-3d-tiles": t("Add OSM 3D Tiles"),
-      "hide-fly-to-delete-layer": t("Hide Fly To Delete Layer"),
-      "override-layer-data": t("Override Layer Data"),
-      "show-selected-features-info": t("Show Selected Features Information"),
-      "layer-styling-examples": t("Layer Styling Examples"),
-      "feature-style-3d-model": t("Feature Style 3D Model"),
-      "feature-style-3d-tiles": t("Feature Style 3D Tiles"),
-      "filter-features-with-style": t("Filter Features by Style"),
-      "override-style": t("Override Style"),
-      "style-with-condition": t("Style With Condition"),
-      "playback-control": t("Playback Control"),
-      "time-driven-features": t("Time Driven Features"),
-      "time-driven-path": t("Time Driven Path"),
-      "theme-selector": t("Theme Selector"),
-      "extension-property": t("Extension Property"),
-      "zoom-in-out": t("Zoom In Out"),
-      "camera-rotation": t("Camera Rotation"),
-      "camera-position": t("Camera Position"),
-      "shared-plugin-id": t("Shared Plugin")
-    };
-  }, [t]);
+  const { categoryTitles, pluginTitles } = useTitles();
 
   const PluginEntryItem: FC<{
     pluginId: string;
