+++ web/src/beta/features/PluginPlayground/Plugins/presets/camera/cameraPosition.ts
@@ -319,5 +319,5 @@ reearth.extension.on('message', (msg) => {
 
 export const cameraPosition: PluginType = {
   id: "camera-position",
-  files: [widgetFile, yamlFile]
+  files: [yamlFile, widgetFile]
 };
