+++ web/src/beta/features/PluginPlayground/Plugins/presets/camera/cameraRotation.ts
@@ -146,5 +146,5 @@ reearth.extension.on("message", (msg) => {
 
 export const cameraRotation: PluginType = {
   id: "camera-rotation",
-  files: [widgetFile, yamlFile]
+  files: [yamlFile, widgetFile]
 };
