+++ web/src/beta/features/PluginPlayground/Plugins/presets/index.ts
@@ -3,11 +3,11 @@ import { PluginType } from "../constants";
 import { cameraPosition } from "./camera/cameraPosition";
 import { cameraRotation } from "./camera/cameraRotation";
 import { zoomInOut } from "./camera/zoomInOut";
-import { extensionExtensionMessenger } from "./communication/extensionExtensionMessenger";
-import { uiExtensionMessenger } from "./communication/uiExtensionMessenger";
 import { myPlugin } from "./custom/myPlugin";
-import { themeSelector } from "./dataStorage/themeSelector";
-import { extensionProperty } from "./extension/extensionProperty";
+import { clientStorageThemeSelector } from "./data/clientStorageThemeSelector";
+import { extensionExtensionMessenger } from "./data/extensionExtensionMessenger";
+import { extensionProperty } from "./data/extensionProperty";
+import { uiExtensionMessenger } from "./data/uiExtensionMessenger";
 import { add3dTiles } from "./layers/add-3Dtiles";
 import { addCsv } from "./layers/add-csv";
 import { addCzml } from "./layers/add-czml";
@@ -19,12 +19,12 @@ import { addWms } from "./layers/add-wms";
 import { hideFlyToDeleteLayer } from "./layers/hideFlyToDeleteLayer";
 import { overrideLayerData } from "./layers/overrideLayerData";
 import { showFeaturesInfo } from "./layers/showSelectedFeaturesInformation";
+import { featureStyle3dModel } from "./layerStyles/featureStyle3dmodel";
+import { featureStyle3dTiles } from "./layerStyles/featureStyle3dTiles";
+import { filterFeatureWithStyle } from "./layerStyles/filterFeaturebyStyle";
 import { layerStylingExamples } from "./layerStyles/layerStylingExamples";
-import { featureStyle3dModel } from "./manageLayerStyle/featureStyle3dmodel";
-import { featureStyle3dTiles } from "./manageLayerStyle/featureStyle3dTiles";
-import { filterFeatureWithStyle } from "./manageLayerStyle/filterFeaturebyStyle";
-import { overrideStyle } from "./manageLayerStyle/overrideStyle";
-import { styleWithCondition } from "./manageLayerStyle/styleWithCondition";
+import { overrideStyle } from "./layerStyles/overrideStyle";
+import { styleWithCondition } from "./layerStyles/styleWithCondition";
 import { playbackControl } from "./timeline/playbackControl";
 import { timeDrivenFeatures } from "./timeline/timeDrivenFeatures";
 import { timeDrivenPath } from "./timeline/timeDrivenPath";
@@ -33,11 +33,11 @@ import { modalWindow } from "./ui/modalWindow";
 import { popupPlugin } from "./ui/popup";
 import { responsivePanel } from "./ui/responsivePanel";
 import { sidebar } from "./ui/sidebar";
-import { enableShadowStyle } from "./viewerAndSceneSettings/enableShadowStyle";
-import { enableTerrain } from "./viewerAndSceneSettings/enableTerrain";
-import { mouseEvents } from "./viewerAndSceneSettings/mouseEvent";
-import { showLabel } from "./viewerAndSceneSettings/showLabel";
-import { takeScreenshot } from "./viewerAndSceneSettings/takeScreenshot";
+import { enableShadowStyle } from "./viewer/enableShadowStyle";
+import { enableTerrain } from "./viewer/enableTerrain";
+import { mouseEvents } from "./viewer/mouseEvent";
+import { showLabel } from "./viewer/showLabel";
+import { takeScreenshot } from "./viewer/takeScreenshot";
 
 type PresetPluginCategory = {
   id: string;
@@ -55,8 +55,13 @@ export const presetPlugins: PresetPlugins = [
     plugins: [responsivePanel, sidebar, header, modalWindow, popupPlugin]
   },
   {
-    id: "communication",
-    plugins: [uiExtensionMessenger, extensionExtensionMessenger]
+    id: "data",
+    plugins: [
+      uiExtensionMessenger,
+      extensionExtensionMessenger,
+      extensionProperty,
+      clientStorageThemeSelector
+    ]
   },
   {
     id: "viewerScene",
@@ -102,13 +107,5 @@ export const presetPlugins: PresetPlugins = [
   {
     id: "timeline",
     plugins: [playbackControl, timeDrivenFeatures, timeDrivenPath]
-  },
-  {
-    id: "dataStorage",
-    plugins: [themeSelector]
-  },
-  {
-    id: "extension",
-    plugins: [extensionProperty]
   }
 ];
