+++ web/src/beta/features/PluginPlayground/Plugins/presets/layerStyles/featureStyle3dTiles.ts
@@ -91,5 +91,5 @@ reearth.camera.flyTo(
 
 export const featureStyle3dTiles: PluginType = {
   id: "feature-style-3d-tiles",
-  files: [widgetFile, yamlFile]
+  files: [yamlFile, widgetFile]
 };
