+++ web/src/beta/features/PluginPlayground/Plugins/presets/layerStyles/featureStyle3dmodel.ts
@@ -123,5 +123,5 @@ reearth.camera.flyTo(
 
 export const featureStyle3dModel: PluginType = {
   id: "feature-style-3d-model",
-  files: [widgetFile, yamlFile]
+  files: [yamlFile, widgetFile]
 };
