+++ web/src/beta/features/PluginPlayground/Plugins/presets/layerStyles/filterFeaturebyStyle.ts
@@ -161,5 +161,5 @@ reearth.extension.on("message", (msg) => {
 
 export const filterFeatureWithStyle: PluginType = {
   id: "filter-features-with-style",
-  files: [widgetFile, yamlFile]
+  files: [yamlFile, widgetFile]
 };
