+++ web/src/beta/features/PluginPlayground/Plugins/presets/layerStyles/styleWithCondition.ts
@@ -95,5 +95,5 @@ reearth.timeline.setTime({
 
 export const styleWithCondition: PluginType = {
   id: "style-with-condition",
-  files: [widgetFile, yamlFile]
+  files: [yamlFile, widgetFile]
 };
