+++ web/src/beta/features/PluginPlayground/Plugins/presets/layers/add-OSM-3DTiles.ts
@@ -60,5 +60,5 @@ reearth.camera.flyTo(
 
 export const addOsm3dTiles: PluginType = {
   id: "add-osm-3d-tiles",
-  files: [widgetFile, yamlFile]
+  files: [yamlFile, widgetFile]
 };
