+++ web/src/beta/features/PluginPlayground/Plugins/presets/layers/add-csv.ts
@@ -42,5 +42,5 @@ reearth.layers.add(layerCsv);
 
 export const addCsv: PluginType = {
   id: "add-csv",
-  files: [widgetFile, yamlFile]
+  files: [yamlFile, widgetFile]
 };
