+++ web/src/beta/features/PluginPlayground/Plugins/presets/layers/add-czml.ts
@@ -96,5 +96,5 @@ reearth.camera.flyTo(
 
 export const addCzml: PluginType = {
   id: "add-czml",
-  files: [widgetFile, yamlFile]
+  files: [yamlFile, widgetFile]
 };
