+++ web/src/beta/features/PluginPlayground/Plugins/presets/layers/add-kml.ts
@@ -76,5 +76,5 @@ reearth.layers.add(layerKmlUrl);`
 
 export const addKml: PluginType = {
   id: "add-kml",
-  files: [widgetFile, yamlFile]
+  files: [yamlFile, widgetFile]
 };
