+++ web/src/beta/features/PluginPlayground/Plugins/presets/layers/add-wms.ts
@@ -39,5 +39,5 @@ reearth.layers.add(layerWmsUrl);
 
 export const addWms: PluginType = {
   id: "add-wms",
-  files: [widgetFile, yamlFile]
+  files: [yamlFile, widgetFile]
 };
