+++ web/src/beta/features/PluginPlayground/Plugins/presets/layers/hideFlyToDeleteLayer.ts
@@ -217,5 +217,5 @@ reearth.extension.on("message", (msg) => {
 
 export const hideFlyToDeleteLayer: PluginType = {
   id: "hide-fly-to-delete-layer",
-  files: [widgetFile, yamlFile]
+  files: [yamlFile, widgetFile]
 };
