+++ web/src/beta/features/PluginPlayground/Plugins/presets/layers/overrideLayerData.ts
@@ -177,5 +177,5 @@ LayerManager.createInitialLayer();
 
 export const overrideLayerData: PluginType = {
   id: "override-layer-data",
-  files: [widgetFile, yamlFile]
+  files: [yamlFile, widgetFile]
 };
