+++ web/src/beta/features/PluginPlayground/Plugins/presets/timeline/playbackControl.ts
@@ -390,5 +390,5 @@ const widgetFile: FileType = {
 
 export const playbackControl: PluginType = {
   id: "playback-control",
-  files: [widgetFile, yamlFile]
+  files: [yamlFile, widgetFile]
 };
