+++ web/src/beta/features/PluginPlayground/Plugins/presets/timeline/timeDrivenPath.ts
@@ -125,5 +125,5 @@ reearth.camera.flyTo(
 
 export const timeDrivenPath: PluginType = {
   id: "time-driven-path",
-  files: [widgetFile, yamlFile]
+  files: [yamlFile, widgetFile]
 };
