+++ web/src/beta/features/PluginPlayground/Plugins/presets/ui/header.ts
@@ -81,5 +81,5 @@ const widgetFile: FileType = {
 
 export const header: PluginType = {
   id: "header",
-  files: [widgetFile, yamlFile]
+  files: [yamlFile, widgetFile]
 };
