+++ web/src/beta/features/PluginPlayground/Plugins/presets/ui/modalWindow.ts
@@ -92,5 +92,5 @@ reearth.extension.on("message", msg => {
 
 export const modalWindow: PluginType = {
   id: "modal-window",
-  files: [widgetFile, yamlFile]
+  files: [yamlFile, widgetFile]
 };
