+++ web/src/beta/features/PluginPlayground/Plugins/presets/ui/sidebar.ts
@@ -208,5 +208,5 @@ const widgetFile: FileType = {
 
 export const sidebar: PluginType = {
   id: "sidebar",
-  files: [widgetFile, yamlFile]
+  files: [yamlFile, widgetFile]
 };
