+++ web/src/beta/features/PluginPlayground/Plugins/presets/viewer/enableShadowStyle.ts
@@ -152,5 +152,5 @@ reearth.extension.on("message", (msg) => {
 
 export const enableShadowStyle: PluginType = {
   id: "enable-shadow-style",
-  files: [widgetFile, yamlFile]
+  files: [yamlFile, widgetFile]
 };
