+++ web/src/beta/features/PluginPlayground/Plugins/presets/viewer/enableTerrain.ts
@@ -125,5 +125,5 @@ reearth.extension.on("message", (msg) => {
 
 export const enableTerrain: PluginType = {
   id: "enable-terrain",
-  files: [widgetFile, yamlFile]
+  files: [yamlFile, widgetFile]
 };
