+++ web/src/beta/features/PluginPlayground/Plugins/presets/viewer/showLabel.ts
@@ -53,5 +53,5 @@ reearth.camera.flyTo(
 
 export const showLabel: PluginType = {
   id: "show-label",
-  files: [widgetFile, yamlFile]
+  files: [yamlFile, widgetFile]
 };
